+++ core/Options.lua
@@ -171,7 +171,7 @@ function Options:Initialize()
     local variableTable = PER.data.options
     local category, layout = Settings.RegisterVerticalLayoutSubcategory(mainCategory, L["options"])
 
-	local ParentSettingRaceTracker
+	local ParentCheckboxRaceTrackerSpeed
 
     layout:AddInitializer(CreateSettingsListSectionHeaderInitializer(L["options.race-tracker"]))
 
@@ -204,43 +204,22 @@ function Options:Initialize()
     end
 
     do
-        local name = L["options.race-tracker-gliding-speed.name"]
-        local tooltip = L["options.race-tracker-gliding-speed.tooltip"]
-        local variable = "race-tracker-gliding-speed"
-        local defaultValue = false
-
-        local setting = Settings.RegisterAddOnSetting(category, addonName .. "_" .. variable, variable, variableTable, Settings.VarType.Boolean, name, defaultValue)
-        Settings.CreateCheckbox(category, setting, tooltip)
-    end
-
-    do
         local name = L["options.race-tracker-background.name"]
         local tooltip = L["options.race-tracker-background.tooltip"]
         local variable = "race-tracker-background"
-        local defaultValue = true
-
-        local setting = Settings.RegisterAddOnSetting(category, variable, variable, variableTable, Settings.VarType.Boolean, name, defaultValue)
-        ParentSettingBackground = Settings.CreateCheckbox(category, setting, tooltip)
-    end
-
-    do
-        local name = L["options.race-tracker-background-type.name"]
-        local tooltip = L["options.race-tracker-background-type.tooltip"]
-        local variable = "race-tracker-background-type"
         local defaultValue = 0
 
         local function GetOptions()
             local container = Settings.CreateControlTextContainer()
-            container:Add(0, L["options.race-tracker-background-type.value.0"])
-            container:Add(1, L["options.race-tracker-background-type.value.1"])
-			container:Add(2, L["options.race-tracker-background-type.value.2"])
+            container:Add(0, L["options.race-tracker-background.value.0"])
+            container:Add(1, L["options.race-tracker-background.value.1"])
+			container:Add(2, L["options.race-tracker-background.value.2"])
+			container:Add(3, L["options.race-tracker-background.value.3"])
             return container:GetData()
         end
 
         local setting = Settings.RegisterAddOnSetting(category, addonName .. "_" .. variable, variable, variableTable, Settings.VarType.Number, name, defaultValue)
-        local subSetting = Settings.CreateDropdown(category, setting, GetOptions, tooltip)
-
-        subSetting:SetParentInitializer(ParentSettingBackground, function() return PER.data.options["race-tracker-background"] end)
+        Settings.CreateDropdown(category, setting, GetOptions, tooltip)
     end
 
     do
@@ -254,9 +233,10 @@ function Options:Initialize()
         local step = 10
 
         local setting = Settings.RegisterAddOnSetting(category, addonName .. "_" .. variable, variable, variableTable, Settings.VarType.Number, name, defaultValue)
-        local options = Settings.CreateSliderOptions(minValue, maxValue, step)
 
+		local options = Settings.CreateSliderOptions(minValue, maxValue, step)
         options:SetLabelFormatter(MinimalSliderWithSteppersMixin.Label.Right)
+
         Settings.CreateSlider(category, setting, options, tooltip)
     end
 
@@ -271,9 +251,10 @@ function Options:Initialize()
         local step = 10
 
         local setting = Settings.RegisterAddOnSetting(category, addonName .. "_" .. variable, variable, variableTable, Settings.VarType.Number, name, defaultValue)
-        local options = Settings.CreateSliderOptions(minValue, maxValue, step)
 
+		local options = Settings.CreateSliderOptions(minValue, maxValue, step)
         options:SetLabelFormatter(MinimalSliderWithSteppersMixin.Label.Right)
+
         Settings.CreateSlider(category, setting, options, tooltip)
     end
 
@@ -288,10 +269,59 @@ function Options:Initialize()
         local step = 1
 
         local setting = Settings.RegisterAddOnSetting(category, addonName .. "_" .. variable, variable, variableTable, Settings.VarType.Number, name, defaultValue)
-        local options = Settings.CreateSliderOptions(minValue, maxValue, step)
 
+		local options = Settings.CreateSliderOptions(minValue, maxValue, step)
         options:SetLabelFormatter(MinimalSliderWithSteppersMixin.Label.Right)
+
         Settings.CreateSlider(category, setting, options, tooltip)
+    end
+
+    do
+        local name = L["options.race-tracker-speed-display.name"]
+        local tooltip = L["options.race-tracker-speed-display.tooltip"]
+        local variable = "race-tracker-speed-display"
+        local defaultValue = false
+
+        local setting = Settings.RegisterAddOnSetting(category, addonName .. "_" .. variable, variable, variableTable, Settings.VarType.Boolean, name, defaultValue)
+        ParentCheckboxRaceTrackerSpeed = Settings.CreateCheckbox(category, setting, tooltip)
+    end
+
+    do
+        local name = L["options.race-tracker-speed-display-horizontal-shift.name"]
+        local tooltip = L["options.race-tracker-speed-display-horizontal-shift.tooltip"]
+        local variable = "race-tracker-speed-display-horizontal-shift"
+        local defaultValue = 0
+
+        local minValue = -500
+        local maxValue = 500
+        local step = 10
+
+        local setting = Settings.RegisterAddOnSetting(category, addonName .. "_" .. variable, variable, variableTable, Settings.VarType.Number, name, defaultValue)
+
+		local options = Settings.CreateSliderOptions(minValue, maxValue, step)
+        options:SetLabelFormatter(MinimalSliderWithSteppersMixin.Label.Right)
+
+        local slider = Settings.CreateSlider(category, setting, options, tooltip)
+		slider:SetParentInitializer(ParentCheckboxRaceTrackerSpeed, function() return PER.data.options["race-tracker-speed-display"] end)
+    end
+
+    do
+        local name = L["options.race-tracker-speed-display-vertical-shift.name"]
+        local tooltip = L["options.race-tracker-speed-display-vertical-shift.tooltip"]
+        local variable = "race-tracker-speed-display-vertical-shift"
+        local defaultValue = -100
+
+        local minValue = -400
+        local maxValue = 400
+        local step = 10
+
+        local setting = Settings.RegisterAddOnSetting(category, addonName .. "_" .. variable, variable, variableTable, Settings.VarType.Number, name, defaultValue)
+
+		local options = Settings.CreateSliderOptions(minValue, maxValue, step)
+        options:SetLabelFormatter(MinimalSliderWithSteppersMixin.Label.Right)
+
+        local slider = Settings.CreateSlider(category, setting, options, tooltip)
+		slider:SetParentInitializer(ParentCheckboxRaceTrackerSpeed, function() return PER.data.options["race-tracker-speed-display"] end)
     end
 
     layout:AddInitializer(CreateSettingsListSectionHeaderInitializer(L["options.race-time-overview"]))