+++ CHANGELOG.md
@@ -3,5 +3,5 @@
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Percursus/releases)
 
 - Added: New Race Tracker backgrounds (Alliance and Horde)
-- Changed: The speed display during a race is shown as a bar
+- Changed: The speed display during a race is shown as a bar and can be positioned separately from the Race Tracker
 - Minor code adjustments